refactor(diagram_analysis): deepen seams + extract analysis_format peer
Implements six structural changes to diagram_analysis/ and promotes the
wire-format concerns into a sibling top-level package.

1. Inline IncrementalRunner + FullAnalysisRunner back into DiagramGenerator.
   Both runners reached through self._gen for ~7 generator privates each
   and had a single call site, so the seam wasn't real. Their bodies now
   live in generate_analysis() / generate_analysis_incremental() and
   _expand_subcomponents().

2. Compute the IncrementalDelta once. The pipeline previously built one
   delta for the wire-format payload and the runner built another with
   component_resolver wired. Pipeline now builds the fully-resolved delta
   and hands it to the generator, which consumes it.

3. Type the OwnershipIndex. Replaces the dict[str, dict] grab-bag with a
   frozen dataclass; pick_component_for_file becomes a method, magic-key
   lookups (ownership_index["..."]) become attribute access in
   derive_patch_scopes and the pipeline-side resolver lambda.

4. Collapse three SymbolResolver implementations to one. Extracts
   StaticAnalysisSymbolResolver (was buried in incremental_pipeline) into
   diagram_analysis/symbol_resolver.py, used by both pipeline and
   generator. Removes DiagramGenerator._collect_method_entries_from_static_analysis
   and ._normalize_repo_path along with the inline lambda.

5. Split incremental_updater.py: delta construction stays
   (IncrementalUpdater + types); apply_method_delta /
   prune_empty_components / drop_deltas_for_pruned_components /
   _sync_component_methods move to diagram_analysis/delta_application.py.

6. Split analysis_json.py along caller-group lines: Pydantic models,
   serializers, and the parser + build_id_to_name_map + (now public)
   compute_depth_level become three modules.

Then promote the format layer to a top-level peer:

7. Move analysis_json_{models,serialize,parse}.py, io_utils.py, and
   ease.py into a new analysis_format/ package. Both the engine and the
   incremental pipeline now depend on it; it depends on neither.
