+++ diagram_analysis/scope_planner.py
@@ -1,6 +1,7 @@
 """Pure helpers for routing trace results into ``PatchScope`` lists."""
 
 from collections import defaultdict
+from dataclasses import dataclass
 from pathlib import Path
 
 from agents.agent_responses import AnalysisInsights
@@ -8,30 +9,70 @@ from diagram_analysis.analysis_patcher import PatchScope, patch_analysis_scope
 from diagram_analysis.incremental_models import TraceResult, TraceStopReason
 
 
+@dataclass(frozen=True)
+class OwnershipIndex:
+    """Routes files and methods to their owning components.
+
+    Deeper-nested components win on ties: when a method appears in multiple
+    components, ``method_to_component`` keeps the deepest one.
+    """
+
+    component_to_scope: dict[str, str | None]
+    file_to_components: dict[str, set[str]]
+    component_to_files: dict[str, set[str]]
+    method_to_component: dict[str, str]
+    component_to_descendants: dict[str, set[str]]
+
+    def pick_component_for_file(
+        self,
+        file_path: str,
+        rename_map: dict[str, str] | None = None,
+    ) -> str | None:
+        if file_path in self.file_to_components:
+            candidates = sorted(self.file_to_components[file_path], key=lambda cid: cid.count("."))
+            return candidates[-1] if candidates else None
+
+        inverse_renames = {new_path: old_path for old_path, new_path in (rename_map or {}).items()}
+        old_path = inverse_renames.get(file_path)
+        if old_path and old_path in self.file_to_components:
+            candidates = sorted(self.file_to_components[old_path], key=lambda cid: cid.count("."))
+            return candidates[-1] if candidates else None
+
+        if not self.file_to_components:
+            return None
+
+        distances = {candidate: directory_distance(file_path, candidate) for candidate in self.file_to_components}
+        nearest_distance = min(distances.values())
+        nearest_files = [
+            candidate
+            for candidate, distance in distances.items()
+            if distance == nearest_distance and candidate != file_path
+        ]
+        nearest_components = sorted(
+            {cid for candidate in nearest_files for cid in self.file_to_components.get(candidate, set())}
+        )
+        if not nearest_components:
+            return None
+        if len(nearest_components) == 1:
+            return nearest_components[0]
+        return lowest_common_ancestor(nearest_components)
+
+    def descendants_of(self, component_id: str) -> set[str]:
+        return self.component_to_descendants.get(component_id, {component_id})
+
+
 def build_ownership_index(
     root_analysis: AnalysisInsights,
     sub_analyses: dict[str, AnalysisInsights],
-) -> dict[str, dict]:
-    """Build maps used for routing files/methods to their owning components."""
+) -> OwnershipIndex:
+    """Build the index used for routing files/methods to their owning components."""
     component_to_scope: dict[str, str | None] = {}
     file_to_components: dict[str, set[str]] = defaultdict(set)
     component_to_files: dict[str, set[str]] = defaultdict(set)
     method_to_component: dict[str, str] = {}
 
-    for component in root_analysis.components:
-        component_to_scope[component.component_id] = None
-        for group in component.file_methods:
-            file_to_components[group.file_path].add(component.component_id)
-            component_to_files[component.component_id].add(group.file_path)
-            for method in group.methods:
-                existing = method_to_component.get(method.qualified_name)
-                if existing is None or component.component_id.count(".") >= existing.count("."):
-                    method_to_component[method.qualified_name] = component.component_id
-
-    for scope_id, analysis in sorted(sub_analyses.items(), key=lambda item: item[0].count(".")):
-        # Route subtree patching to the nested scope when a component has a dedicated sub-analysis.
-        component_to_scope[scope_id] = scope_id
-        for component in analysis.components:
+    def absorb(scope_id: str | None, components) -> None:
+        for component in components:
             component_to_scope[component.component_id] = scope_id
             for group in component.file_methods:
                 file_to_components[group.file_path].add(component.component_id)
@@ -41,21 +82,29 @@ def build_ownership_index(
                     if existing is None or component.component_id.count(".") >= existing.count("."):
                         method_to_component[method.qualified_name] = component.component_id
 
-    component_to_descendants: dict[str, set[str]] = {}
-    for component_id in component_to_scope:
-        component_to_descendants[component_id] = {
+    absorb(None, root_analysis.components)
+
+    for scope_id, analysis in sorted(sub_analyses.items(), key=lambda item: item[0].count(".")):
+        # Route subtree patching to the nested scope when a component has a dedicated sub-analysis.
+        component_to_scope[scope_id] = scope_id
+        absorb(scope_id, analysis.components)
+
+    component_to_descendants: dict[str, set[str]] = {
+        component_id: {
             other_id
             for other_id in component_to_scope
             if other_id == component_id or other_id.startswith(component_id + ".")
         }
-
-    return {
-        "component_to_scope": component_to_scope,
-        "file_to_components": file_to_components,
-        "component_to_files": component_to_files,
-        "method_to_component": method_to_component,
-        "component_to_descendants": component_to_descendants,
+        for component_id in component_to_scope
     }
+
+    return OwnershipIndex(
+        component_to_scope=component_to_scope,
+        file_to_components=dict(file_to_components),
+        component_to_files=dict(component_to_files),
+        method_to_component=method_to_component,
+        component_to_descendants=component_to_descendants,
+    )
 
 
 def _scope_component_ids(analysis: AnalysisInsights) -> set[str]:
@@ -87,43 +136,6 @@ def directory_distance(left_path: str, right_path: str) -> int:
     return (len(left_parts) - common) + (len(right_parts) - common)
 
 
-def pick_component_for_file(
-    file_path: str,
-    ownership_index: dict[str, dict],
-    rename_map: dict[str, str] | None = None,
-) -> str | None:
-    file_to_components: dict[str, set[str]] = ownership_index["file_to_components"]
-
-    if file_path in file_to_components:
-        candidates = sorted(file_to_components[file_path], key=lambda component_id: component_id.count("."))
-        return candidates[-1] if candidates else None
-
-    inverse_renames = {new_path: old_path for old_path, new_path in (rename_map or {}).items()}
-    old_path = inverse_renames.get(file_path)
-    if old_path and old_path in file_to_components:
-        candidates = sorted(file_to_components[old_path], key=lambda component_id: component_id.count("."))
-        return candidates[-1] if candidates else None
-
-    if not file_to_components:
-        return None
-
-    distances = {candidate: directory_distance(file_path, candidate) for candidate in file_to_components}
-    nearest_distance = min(distances.values())
-    nearest_files = [
-        candidate
-        for candidate, distance in distances.items()
-        if distance == nearest_distance and candidate != file_path
-    ]
-    nearest_components = sorted(
-        {component_id for candidate in nearest_files for component_id in file_to_components.get(candidate, set())}
-    )
-    if not nearest_components:
-        return None
-    if len(nearest_components) == 1:
-        return nearest_components[0]
-    return lowest_common_ancestor(nearest_components)
-
-
 def normalize_changes_for_delta(changes) -> tuple[list[str], list[str], list[str], dict[str, str]]:
     added_files = set(changes.added_files)
     modified_files = set(changes.modified_files)
@@ -146,19 +158,17 @@ def derive_patch_scopes(
     trace_result: TraceResult,
     root_analysis: AnalysisInsights,
     sub_analyses: dict[str, AnalysisInsights],
-    ownership_index: dict[str, dict],
+    ownership_index: OwnershipIndex,
     rename_map: dict[str, str] | None = None,
 ) -> list[PatchScope]:
-    component_to_scope: dict[str, str | None] = ownership_index["component_to_scope"]
-    component_to_descendants: dict[str, set[str]] = ownership_index["component_to_descendants"]
-    method_to_component: dict[str, str] = ownership_index["method_to_component"]
+    method_to_component = ownership_index.method_to_component
 
     target_component_ids: set[str] = {
         method_to_component[method] for method in trace_result.visited_methods if method in method_to_component
     }
 
     for file_path in trace_result.non_traceable_files + trace_result.disconnected_files:
-        component_id = pick_component_for_file(file_path, ownership_index, rename_map)
+        component_id = ownership_index.pick_component_for_file(file_path, rename_map)
         if component_id:
             target_component_ids.add(component_id)
 
@@ -170,7 +180,7 @@ def derive_patch_scopes(
 
     target_by_scope: dict[str | None, set[str]] = defaultdict(set)
     for component_id in target_component_ids:
-        target_by_scope[component_to_scope.get(component_id)].add(component_id)
+        target_by_scope[ownership_index.component_to_scope.get(component_id)].add(component_id)
 
     should_widen = trace_result.stop_reason in {
         TraceStopReason.UNCERTAIN,
@@ -191,7 +201,7 @@ def derive_patch_scopes(
             if lca:
                 descendant_ids = {
                     component_id
-                    for component_id in component_to_descendants.get(lca, {lca})
+                    for component_id in ownership_index.descendants_of(lca)
                     if component_id in scope_component_ids
                 }
                 if descendant_ids:
@@ -218,7 +228,7 @@ def derive_patch_scopes(
                 synthetic_files=[
                     file_path
                     for file_path in trace_result.non_traceable_files + trace_result.disconnected_files
-                    if pick_component_for_file(file_path, ownership_index, rename_map) in selected_ids
+                    if ownership_index.pick_component_for_file(file_path, rename_map) in selected_ids
                 ],
                 semantic_impact_summary=trace_result.semantic_impact_summary,
             )
