+++ pyproject.toml
@@ -77,6 +77,7 @@ packages = [
     "agents",
     "agents.prompts",
     "agents.tools",
+    "analysis_format",
     "core",
     "diagram_analysis",
     "health",
